requirements.txt
PyPDF2
python-pptx
moviepy
pytube
imageio
imageio-ffmpeg